Nope - still don't understand how python modules work.
So in order to have an lsst.sqre hierarchy, if you declare in setup.py
the package as lsst.sqre, no __init.py__ gets copied into the lsst site
packages directory and the import fails. Let's see if lsst is sufficient
or whether lsst.sqre needs to be added to the dict.