cycles() returns a list, not a generator.
The number of cycles is tiny, so we're not saving any worthwhile memory by
using a generator expression. Plus, it's convenient to be able to look things
up based on their order in the sequence -- jirakit does this -- or to be able
to iterate over it multiple times -- jira2dot does this.