jira2dot & jira2txt return strings.
Rather than accepting as an input some sort of buffer and attempting to write
to it.

This makes it easier to maintain compatibility with both Python 2 & Python 3,
as we don't need to worry about whether we have a buffer of bytes or a buffer
of unicode and behave appropriately within the generating routine; all
conversions can be handled in a single place in the dlp script.